Add files via upload
Traffic Stop (by police) Dataset
This dataset contains detailed information about traffic stops conducted by law enforcement. It includes various attributes such as:

stop_date: The date on which the traffic stop occurred.
stop_time: The time at which the stop was initiated.
driver_gender: The gender of the driver.
driver_age: The age of the driver at the time of the stop.
driver_race: The race/ethnicity of the driver.
violation: The type of violation for which the stop was made (e.g., speeding, equipment violation).
search_conducted: Indicates whether a vehicle search was conducted during the stop.
search_type: The type of search performed, if applicable.
stop_outcome: The outcome of the stop (e.g., citation, arrest).
is_arrested: Whether or not the driver was arrested.
stop_duration: The length of time the stop lasted.
drugs_related_stop: Indicates whether the stop was related to drugs.
The dataset provides a comprehensive overview of traffic stops, including demographic information about drivers, the nature of the violations, and the outcomes of the stops, allowing for in-depth analysis of law enforcement practices.